refactor: move config to theme dir
